Fix:gui_internal:Fix signed integer overflow in gui_internal_gesture_ring_add.|Part of fix for #1145.
git-svn-id: http://svn.code.sf.net/p/navit/code/trunk/navit@5546 ffa7fe5e-494d-0410-b361-a75ebd5db220

+++ navit/gui/internal/gui_internal_priv.h
@@ -147,7 +147,7 @@ struct gui_priv {
 /* gestures */	
 
 	struct gesture_elem {
-		int msec;
+		long long msec;
 		struct point p;
 	} gesture_ring[GESTURE_RINGSIZE];
 	int gesture_ring_last, gesture_ring_first;
